freeMemCount is not the free page count
The total pages need to be counted in the linked list. That would give the free page count, and not the variable, freeMemCount, which only calculated the free pages within the range between *vstart and *vend. While printing, multiply with PGSIZE to get the output in the form of bytes and not in the form of number of pages, as earlier.